requirements.txt
openai>=1.30.0
aiohttp>=3.9.0
requests>=2.31.0
python-dotenv>=1.0.0
